requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
twilio==9.4.4
python-multipart==0.0.20
